# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/accounts/admin/{tenant}/devices": {
            "get": {
                "tags": [
                    "device"
                ],
                "summary": "list device",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "device"
                ],
                "summary": "new device",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/code/{userCode}": {
            "post": {
                "tags": [
                    "device"
                ],
                "summary": "delete device groups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tenant",
                        "name": "userCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/{deviceId}": {
            "get": {
                "tags": [
                    "device"
                ],
                "summary": "get device",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "device"
                ],
                "summary": "update device",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "device"
                ],
                "summary": "delete device",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/{deviceId}/groups": {
            "get": {
                "tags": [
                    "device"
                ],
                "summary": "list device groups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "device"
                ],
                "summary": "new device groups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/{deviceId}/groups/{groupId}": {
            "put": {
                "tags": [
                    "device"
                ],
                "summary": "update device groups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "device"
                ],
                "summary": "delete device groups",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/{deviceId}/secrets": {
            "get": {
                "tags": [
                    "client"
                ],
                "summary": "get client secrets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "client"
                ],
                "summary": "get client secrets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/devices/{deviceId}/secrets/{secretId}": {
            "delete": {
                "tags": [
                    "client"
                ],
                "summary": "get client secrets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "tenant",
                        "name": "secretId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}